moved string functions.
